Migrate to android openssl git repo instead of old guardian project's one.
git-svn-id: http://csipsimple.googlecode.com/svn/trunk/CSipSimple@2415 9f815046-5998-e9c0-b7e2-ac03a23edfa4

+++ jni/csipsimple-wrapper/src/pjsip_opus_sdp_rewriter.c
@@ -106,7 +106,7 @@ static pj_status_t opus_sdp_on_tx_msg(pjsip_tx_data *tdata)
                 new_value.ptr = (char*) pj_pool_alloc(tdata->pool, 20);
                 new_value.slen = pj_ansi_snprintf(new_value.ptr, 20, "%.*s%.*s",
                                                                                                                 found_opus - opus_attr->value.ptr, opus_attr->value.ptr,
-                                                                                                                STR_OPUS_RFC_FMT.slen, STR_OPUS_RFC_FMT.ptr);
+                                                                                                                (int)STR_OPUS_RFC_FMT.slen, STR_OPUS_RFC_FMT.ptr);
                 opus_attr->value = new_value;
                 tdata->msg->body = new_body;
             }
